Fix example plotting bugs exposed by run-examples CI
- na_filtering_part_{1,2,3}: cast f_max to float (was array from k_max)
- tvsp_steering_linear_array: infer p_final grid shape instead of
  hardcoding 128x128 (PML stripping changes the size)
- ivp_loading_external_image: fix _REPO_ROOT after move from
  examples/ported/ to examples/ (one fewer parent dir)

+++ examples/na_filtering_part_1.py
@@ -132,7 +132,7 @@ if __name__ == "__main__":
     input_as = np.abs(np.fft.rfft(source_p)) / N
     output_as = np.abs(np.fft.rfft(p[sensor_row, :])) / N
 
-    f_max = kgrid.k_max * np.min(medium.sound_speed) / (2 * np.pi)
+    f_max = float(kgrid.k_max * np.min(medium.sound_speed) / (2 * np.pi))
     f_MHz = f * 1e-6
 
     ax2.plot(f_MHz, input_as, "k-", label="Input spectrum")
