+++ examples/na_filtering_part_2.py
@@ -151,7 +151,7 @@ if __name__ == "__main__":
     f = np.fft.rfftfreq(N, d=dt)
     output_as = np.abs(np.fft.rfft(p[sensor_row, :])) / N
 
-    f_max = kgrid.k_max * np.min(medium.sound_speed) / (2 * np.pi)
+    f_max = float(kgrid.k_max * np.min(medium.sound_speed) / (2 * np.pi))
     f_MHz = f * 1e-6
 
     ax2.plot(f_MHz, output_as, "b-", label="Recorded spectrum")
